# openapi/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: AWS Environment Ondemand APIs
  description: |-
    This is an end to end environment ondemand setup api
  contact:
    email: apiteam@swagger.io
  version: 1.0.11
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /environment:
    get:
      tags:
        - Environment
      summary: Get all the environments
      description: List all the components.
      operationId: getAllEnvironments
      parameters:
        - name: environments
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Environment'
        '400':
          description: Invalid app
    post:
      tags:
        - Environment
      summary: Create a new environment on aws
      description: Create a new environment on aws
      operationId: createEnvironment
      requestBody:
        description: Create a new environment in the aws
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Environment'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    put:
      tags:
        - Environment
      summary: Update an existing environment
      description: Update an existing environment by Id
      operationId: updateEnvironment
      requestBody:
        description: Update an existent environment in the aws
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Environment'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '422':
          description: Validation exception
  /environment/{envId}:
    get:
      tags:
        - Environment
      summary: Find environment by ID
      description: Returns a single environment
      operationId: getEnvById
      parameters:
        - name: envId
          in: path
          description: ID of environment to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          description: Invalid ID supplied
        '404':
          description: Environments not found
    delete:
      tags:
        - Environment
      summary: Deletes an environment
      description: delete a environment
      operationId: deleteEnv
      parameters:
        - name: envId
          in: path
          description: Environment id to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid environment value
  /environment/destroy:
    post:
      tags:
        - Environment
      summary: Destroy an environment
      description: Destroy an environment
      operationId: destroyEnvironment
      requestBody:
        description: Destroy an application in the aws
        content:
          application/json:
            example: '{"id" : "uuid"}'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
    get:
      tags:
        - Environment
      summary: Get all the destroyed environments
      description: List all the components.
      operationId: getAllDestroyedEnvs
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Environment'
        '400':
          description: Invalid input
components:
  schemas:
...
    Environment:
      required:
        - appRelease
        - awsAccount
        - namespace
      type: object
      properties:
        id:
          type: string
          example: 'uuid'
        appRelease:
          type: string
          example: 'xyz-1.1'
        awsAccount:
          type: string
          example: 'abc'
        namespace:
          type: string
          example: 'xyz-ns'
      xml:
        name: environment
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'